Initial by hand prototyping
I first tried an approach in which the rows were blocks, and the
squares were inline blocks. I had problems with excess space between
the blocks, even with 0 margin. Horizontally, this was caused by
excess space in the source HTML. That is, even minimal
whitespace. This can be solved by squishing together the HTML, at the
cost of readability. But it doesn't address the excess vertical
spacing. I think this can be solved with negative margins, but that's
hacky.

Next, I tried using flex, with the outer container a flex box oriented
column-wise, containing the rows. Then each row is a flex box oriented
row-wise, containing the squares. This has no problem with excess
space, and is what I will go with.

I should perhaps check out grid layouts as well, but that's future work.