Add drawing modes to lighten and darken
Divide black to white space into 10 grey levels
Add or subtract based on mode

Need to read RGB value, which seems to be a mild hack:
https://stackoverflow.com/a/66623849/9797192
That is, it's string parsing, not actually some kind of object

Note that this isn't precise, since each differential requires tenths
of values, but actual RGB values are integers. But it's good enough.
Alpha is ignored.

Darken and lighten are bounded at the extremes.

If we start out as some non-grey color (which can't yet happen), then
the first darken means 1 level darker than white, and the first
lighten means 1 level lighter than black.