Update param handling wrt event listeners
We don't need to pass the event as a parameter, that's implied.
So these are equivalent:
  someObject.addEventListener('action', (event) => someFunction(event); });
  someObject.addEventListener('action', someFunction);

And if this is all we're passing, we can then do:
  function someFunction(event) { }
But if we don't need the event, this is fine:
  function someFunction() { }

I'm a bit confused by the use of "this" within the function.

Previously I was using it for the target of the listener, but that
doesn't seem to work when passing the event. It seems cleaner to get
the target via `event.currentTarget`. For more notes see:
https://developer.mozilla.org/en-US/docs/Web/API/EventTarget/addEventListener#the_value_of_this_within_the_handler
https://developer.mozilla.org/en-US/docs/Web/API/Event#instance_properties
https://developer.mozilla.org/en-US/docs/Web/API/Event/currentTarget

This still doesn't solve the shift key reload problem